+++ Classes/Bootstrap/Base.php
@@ -38,6 +38,15 @@ abstract class Base
     }
 
     /**
+     * Return the extension key in lower case without underscores
+     * @return string
+     */
+    protected function getExtensionKeyForFlexforms(): string
+    {
+        return strtolower($this->getExtensionKeyAsNamespace());
+    }
+
+    /**
      * Return the LLL path as string
      * @param string $key
      * @return string
@@ -45,5 +54,15 @@ abstract class Base
     protected function getLLL(string $key): string
     {
         return 'LLL:EXT:' . $this->extensionKey . '/Resources/Private/Language/' . $key;
+    }
+
+    /**
+     * Return the path to FlexForm file
+     * @param string $filename
+     * @return string
+     */
+    protected function getFlexformPath(string $filename): string
+    {
+        return 'FILE:EXT:' . $this->extensionKey . '/Configuration/Flexforms/' . $filename;
     }
 }