+++ Classes/Bootstrap/ExtLocalconf.php
@@ -20,6 +20,7 @@ namespace Wacon\Filetransfer\Bootstrap;
 use TYPO3\CMS\Extbase\Utility\ExtensionUtility;
 use Wacon\Filetransfer\Controller\DownloadController;
 use Wacon\Filetransfer\Controller\UploadController;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 class ExtLocalconf extends Base
 {
@@ -29,6 +30,7 @@ class ExtLocalconf extends Base
     public function invoke()
     {
         $this->configurePlugins();
+        $this->addFluidEmailTemplatePath();
     }
 
     /**
@@ -49,13 +51,26 @@ class ExtLocalconf extends Base
 
         ExtensionUtility::configurePlugin(
             $this->getExtensionKeyAsNamespace(),
-            'Upload',
+            'Download',
             [
-                DownloadController::class => 'download',
+                DownloadController::class => 'download,error,expired',
             ],
             [
                 DownloadController::class => 'download',
             ]
         );
+    }
+
+    /**
+     * Add fluid template path into $GLOBALS['TYPO3_CONF_VARS']['MAIL']['templateRootPaths']
+     * @return void
+     */
+    private function addFluidEmailTemplatePath()
+    {
+        $newPath = 'EXT:' . $this->extensionKey . '/Resources/Private/Templates/Email/';
+
+        if (!in_array($newPath, $GLOBALS['TYPO3_CONF_VARS']['MAIL']['templateRootPaths'])) {
+            $GLOBALS['TYPO3_CONF_VARS']['MAIL']['templateRootPaths'][] = $newPath;
+        }
     }
 }