+++ CHANGES.md
@@ -1,3 +1,9 @@
+## Version 0.1.0
+- [IMPORTANT] Implement Download function
+- [FEATURE] Add FlexForm for upload plugin to choose download page
+- [FEATURE] tx_filetransfer_domain_model_upload is adminOnly, Exclude token field
+- [NOTICE] ext_emconf: Wrong Email and company
+
 ## Version 0.0.6
 - [FEATURE] Create FileUploadService to handle the file upload